taskTracker project is added.
TaskTracker is an application in which you can track the task assigned to you by your team leader. Developing a software process usually uses the scrum-agile working principle. This application aims to ease your scrum work. 
Vaadin framework is required to complete this code but now we can only run this code in our local setup.